+++ examples/Converter/CommonCkptTensorReaderDemo.cs
@@ -297,7 +297,7 @@ namespace Converter
 				Native.gguf_write_to_file(g_ctx, outputFileName, true);
 
 				byte[] bytes = File.ReadAllBytes(outputFileName);
-				long totalSize = 0;
+				ulong totalSize = 0;
 				for (int i = 0; i < (int)g_ctx->header.n_tensors; ++i)
 				{
 					gguf_tensor_info* info = &g_ctx->infos[i];
@@ -305,13 +305,13 @@ namespace Converter
 					Console.WriteLine($"{name} is doing, current total byte is {totalSize}");
 
 					CommonTensor tensor = tensors.Find(x => x.Name == name);
-					long size = Math.Max(info->size, (int)g_ctx->alignment);
-					long size_pad = Native.GGML_PAD((int)size, (int)g_ctx->alignment);
+					ulong size = Math.Max(info->size, g_ctx->alignment);
+					ulong size_pad = (ulong)Native.GGML_PAD((int)size, (int)g_ctx->alignment);
 					byte[] data = ReadByteFromFile(tensor);
 					totalSize = totalSize + size_pad;
 					if (size_pad != size)
 					{
-						for (long j = 0; j < size_pad - size; ++j)
+						for (ulong j = 0; j < size_pad - size; ++j)
 						{
 							data = data.Concat(new byte[] { 0 }).ToArray();
 						}
